+++ plugins/polio/tests/test_base.py
@@ -17,13 +17,13 @@ from iaso import models as m
 from iaso.models import Account
 from iaso.test import APITestCase, TestCase
 from plugins.polio.tasks.weekly_email import send_notification_email
-from ..api import CACHE_VERSION
+from ..api.common import CACHE_VERSION
 from ..export_utils import format_date
 from ..models import Config, Round, RoundScope
 from ..preparedness.calculator import get_preparedness_score
 from ..preparedness.exceptions import InvalidFormatError
 from ..preparedness.spreadsheet_manager import *
-from ..api.shared_serializers import CampaignSerializer
+from ..api.campaigns import CampaignSerializer
 
 
 class PolioAPITestCase(APITestCase):
@@ -91,7 +91,7 @@ class PolioAPITestCase(APITestCase):
         self.client = APIClient()
         self.client.force_authenticate(self.yoda)
 
-    @mock.patch("plugins.polio.serializers.SpreadSheetImport")
+    @mock.patch("plugins.polio.api.campaigns.SpreadSheetImport")
     def test_preview_invalid_document(self, mock_SpreadSheetImport, *_):
         mock_SpreadSheetImport.create_for_url.return_value = mock.MagicMock()
         url = "https://docs.google.com/spreadsheets/d/1"
